requirements.txt
aliyun-python-sdk-core-v3==2.9.4
certifi==2019.6.16
chardet==3.0.4
click==7.0
dominate==2.4.0
filetype==1.0.5
flask-bootstrap==3.3.7.1
flask-login==0.4.1
flask-sqlalchemy==2.4.0
flask-wtf==0.14.2
flask==1.1.1
idna==2.8
itsdangerous==1.1.0
jinja2==2.10.1
jmespath==0.9.4
markupsafe==1.1.1
pillow==6.1.0
pycrypto==2.6.1
pycryptodome==3.8.2
pycryptodomex==3.7.2
pymysql==0.9.3
pypdf2==1.26.0
python-alipay-sdk==1.10.1
qrcode==6.1
redis==3.3.8
requests==2.22.0
rq==1.1.0
six==1.12.0
sqlalchemy==1.3.7
urllib3==1.25.3
visitor==0.1.3
werkzeug==0.15.5
wtforms==2.2.1
mysql-connector==2.2.9
gunicorn
